requirements.txt
boto3==1.9.115
setuptools==40.2.0
requests==2.21.0
requirements_parser==0.2.0
